Refactor visualization/plotting.py into submodules
